# sets/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Set:
    description: Set defines the structure for an API set
    properties:
      id:
        description: the id for this set
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      intensity:
        format: double
        type: number
        x-go-name: Intensity
      movement:
        type: string
        x-go-name: Movement
      volume:
        format: double
        type: number
        x-go-name: Volume
    required:
    - id
    type: object
    x-go-package: github.com/hrand1005/training-notebook/sets/data
info:
  description: Documentation for Set API
  title: of Set API
  version: 1.0.0
paths:
  /sets:
    get:
      operationId: readAllSets
      responses:
        "200":
          $ref: '#/responses/setsResponse'
      summary: Read all sets.
      tags:
      - sets
    post:
      operationId: createSet
      responses:
        "201":
          $ref: '#/responses/setResponse'
      summary: Creates a set.
      tags:
      - sets
  /sets/{id}:
    delete:
      operationId: deleteSet
      parameters:
      - description: The id of the set
        format: int64
        in: query
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "204":
          $ref: '#/responses/noContent'
      summary: Delete a set.
      tags:
      - sets
    get:
      operationId: readSet
      parameters:
      - description: The id of the set
        format: int64
        in: query
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/setResponse'
      summary: Read a set.
      tags:
      - sets
    put:
      operationId: updateSet
      parameters:
      - description: The id of the set
        format: int64
        in: query
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/setResponse'
      summary: Update a set.
      tags:
      - sets
produces:
- application/json
responses:
  setResponse:
    description: returns a set in the response
    schema:
      $ref: '#/definitions/Set'
  setsResponse:
    description: returns sets in the response
    schema:
      items:
        $ref: '#/definitions/Set'
      type: array
schemes:
- http
swagger: "2.0"